Move endline source files to new package
