+++ packages/endline/src/bin/endline.ts
