+++ packages/endline/src/endline.ts
