+++ packages/endline/types/index.d.ts
@@ -1,4 +1,4 @@
 /// <reference path="./router.d.ts" />
-import endline from '../dist/endline'
+import endline from '../../../dist/endline'
 
 export default endline
