+++ packages/endline/types/router.d.ts
@@ -1,0 +1,3 @@
+import { Router } from '../../../dist/server/router'
+
+export default Router
