--- types/router.d.ts
@@ -1,3 +1,0 @@
-import { Router } from '../dist/server/router'
-
-export default Router
